requirements.txt
azure-ai-documentintelligence
azure-ai-contentsafety
azure-core
python-dotenv
reportlab